Moved sphere mapping to separate file
